CP: Added basic algo 0.1

+++ control_plane/types.hpp
@@ -6,7 +6,7 @@
 
 typedef uint8_t working_copy_t;
 typedef uint16_t rwnd_t;
-typedef uint32_t qdepth_t;
+typedef uint64_t qdepth_t;
 typedef uint16_t port_t;
 
 #endif
